refactor: move the function
