+++ tool/internal_pkg/generator/custom_template.go
@@ -24,7 +24,6 @@ import (
 	"gopkg.in/yaml.v3"
 
 	"github.com/cloudwego/kitex/tool/internal_pkg/log"
-	"github.com/cloudwego/kitex/tool/internal_pkg/tpl"
 	"github.com/cloudwego/kitex/tool/internal_pkg/util"
 )
 
@@ -196,25 +195,6 @@ func (g *generator) GenerateCustomPackage(pkg *PackageInfo) (fs []*File, err err
 			fs = append(fs, f...)
 		}
 	}
-
-	return fs, nil
-}
-
-func (g *generator) GenerateServerWithMultiService(pkg *PackageInfo) (fs []*File, err error) {
-	g.setImports(ServerWithMultiServiceFileName, pkg)
-	task := &Task{
-		Name: ServerWithMultiServiceFileName,
-		Path: util.JoinPath(g.OutputPath, util.CombineOutputPath(g.GenPath, pkg.Namespace), ServerWithMultiServiceFileName),
-		Text: tpl.ServiceWithMultiServiceTpl,
-	}
-	handle := func(task *Task, pkg *PackageInfo) (*File, error) {
-		return task.Render(pkg)
-	}
-	f, err := g.chainMWs(handle)(task, pkg)
-	if err != nil {
-		return nil, err
-	}
-	fs = append(fs, f)
 
 	return fs, nil
 }
